requirements.txt
fastapi
uvicorn
requests
langchain
groqcloud
langchain_groq
python-dotenv
chromadb
streamlit
pandas